chatbot/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
chromadb>=0.4.22
httpx>=0.26.0
pyyaml>=6.0
pydantic>=2.5.0
